fix: update agent-queries
